update license and add travis.yml

+++ src/main/java/org/bardframework/time/JalaliYearMonth.java
@@ -33,7 +33,6 @@ import static java.time.temporal.ChronoUnit.*;
  * However, any application that makes use of historical dates, and requires them
  * to be accurate will find the ISO-8601 approach unsuitable.
  * <p>
- * <p>
  * This is a <a href="{@docRoot}/java/lang/doc-files/ValueBased.html">value-based</a>
  * class; use of identity-sensitive operations (including reference equality
  * ({@code ==}), identity hash code, or synchronization) on instances of
@@ -41,8 +40,7 @@ import static java.time.temporal.ChronoUnit.*;
  * The {@code equals} method should be used for comparisons.
  *
  * @author Vahid Zafari
- * @implSpec This class is immutable and thread-safe.
- * @since 1.0
+ * This class is immutable and thread-safe.
  */
 public final class JalaliYearMonth implements Temporal, TemporalAdjuster, Comparable<JalaliYearMonth>, Serializable {
 
@@ -747,11 +745,11 @@ public final class JalaliYearMonth implements Temporal, TemporalAdjuster, Compar
                 case YEARS:
                     return plusYears(amountToAdd);
                 case DECADES:
-                    return plusYears(Math.multiplyExact(amountToAdd, 10));
+                    return plusYears(Math.multiplyExact(amountToAdd, 10L));
                 case CENTURIES:
-                    return plusYears(Math.multiplyExact(amountToAdd, 100));
+                    return plusYears(Math.multiplyExact(amountToAdd, 100L));
                 case MILLENNIA:
-                    return plusYears(Math.multiplyExact(amountToAdd, 1000));
+                    return plusYears(Math.multiplyExact(amountToAdd, 1000L));
                 case ERAS:
                     return with(ERA, Math.addExact(getLong(ERA), amountToAdd));
             }
@@ -792,9 +790,9 @@ public final class JalaliYearMonth implements Temporal, TemporalAdjuster, Compar
         }
         long monthCount = year * 12L + (month - 1);
         long calcMonths = monthCount + monthsToAdd;  // safe overflow
-        int newYear = YEAR.checkValidIntValue(Math.floorDiv(calcMonths, 12));
-        int newMonth = Math.floorMod(calcMonths, 12) + 1;
-        return with(newYear, newMonth);
+        int newYear = YEAR.checkValidIntValue(Math.floorDiv(calcMonths, 12L));
+        long newMonth = Math.floorMod(calcMonths, 12L) + 1;
+        return with(newYear, (int) newMonth);
     }
 
     //-----------------------------------------------------------------------
